feat: MiniGameManager 생성
